Some table styles fixes
